Use SvgOutput in the LegendRenderer
This means that LegendRenderer does not use SVG or D3 at all and only rely
on the output implementation.

This change should not change the output or behaviour of LegendRenderer in any
way, except for one this:
the 'xmlns' attribute is now set every time an image is added to the svg legend,
and not only with remote legends.